Implemented login and signup
Modified login and sign up views, enabled authentication with Firebase and used Toasty library for feedback
