+++ app/build.gradle
@@ -44,6 +44,7 @@ android {
 dependencies {
     implementation platform('com.google.firebase:firebase-bom:32.1.0')
     implementation 'com.google.firebase:firebase-messaging-ktx'
+    implementation 'com.google.firebase:firebase-auth-ktx:21.1.0'
 
     def room_version = "2.5.1"
 
@@ -69,4 +70,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     // test
     testImplementation "com.google.truth:truth:1.1.2"
+
+    //    Toasty
+    implementation 'com.github.GrenderG:Toasty:1.5.0'
 }
