--- app/src/main/java/com/example/motomoapp/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.example.motomoapp
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        val intent = Intent(this, MenuInicioActivity::class.java)
-        startActivity(intent)
-    }
-
-
-}
