--- app/src/main/java/com/example/motomoapp/myGiftCardsFragment.kt
@@ -1,4 +1,0 @@
-package com.example.motomoapp
-
-class myGiftCardsFragment {
-}
