+++ app/src/main/java/com/example/motomoapp/view/CartSummaryActivity.kt
@@ -1,8 +1,6 @@
-package com.example.motomoapp
+package com.example.motomoapp.view
 
-import android.content.Context
 import android.content.Intent
-import android.content.SharedPreferences
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.transition.Slide
@@ -11,10 +9,10 @@ import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.widget.Toolbar
-import androidx.core.content.ContextCompat.startActivity
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.motomoapp.R
 import com.example.motomoapp.adapters.CartRecyclerAdapter
 import com.example.motomoapp.databinding.ActivityCartSummaryBinding
 import com.example.motomoapp.models.Carrito
