+++ app/src/main/java/com/example/motomoapp/view/CashPaymentActivity.kt
@@ -1,8 +1,9 @@
-package com.example.motomoapp
+package com.example.motomoapp.view
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.motomoapp.R
 import com.example.motomoapp.models.Carrito
 import com.example.motomoapp.utils.OrderNotification
 import com.example.motomoapp.utils.ReceiverNotification
