+++ app/src/main/java/com/example/motomoapp/view/CreditCardActivity.kt
@@ -1,17 +1,13 @@
-package com.example.motomoapp
+package com.example.motomoapp.view
 
 
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.example.motomoapp.CashActivity
 import com.example.motomoapp.R
 import com.example.motomoapp.databinding.ActivityCreditCardBinding
-import com.example.motomoapp.databinding.ActivityOrderBinding
 import com.example.motomoapp.models.CreditCard
 
 //formulario de registro de tarjeta de credito
