+++ app/src/main/java/com/example/motomoapp/view/GiftCardActivity.kt
@@ -1,4 +1,4 @@
-package com.example.motomoapp
+package com.example.motomoapp.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -10,6 +10,7 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.widget.Toolbar
+import com.example.motomoapp.R
 import com.example.motomoapp.models.GiftCard
 
 class GiftCardActivity : AppCompatActivity() {
