+++ app/src/main/java/com/example/motomoapp/view/SelectPaymentMethodActivity.kt
@@ -1,25 +1,29 @@
-package com.example.motomoapp
+package com.example.motomoapp.view
 
-import android.os.Bundle
 import android.content.Intent
-import android.view.MenuItem
-import androidx.appcompat.app.ActionBarDrawerToggle
-import androidx.appcompat.widget.Toolbar
 import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.MenuItem
+import android.widget.Button
+import androidx.appcompat.widget.Toolbar
+import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.drawerlayout.widget.DrawerLayout
-import com.example.motomoapp.databinding.ActivityMyCreditCartsBinding
-import com.example.motomoapp.models.MyGiftCards
+import com.example.motomoapp.R
 import com.google.android.material.navigation.NavigationView
+import com.example.motomoapp.databinding.ActivitySelectPaymentMethodBinding
+import com.example.motomoapp.models.MyGiftCards
 
-class MyCreditCards : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
+class SelectPaymentMethodActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
 
-    //Variables Globales
+    private lateinit var binding: ActivitySelectPaymentMethodBinding
 
-    private lateinit var binding: ActivityMyCreditCartsBinding
-    private lateinit var creditCardFragment: myCreditCardsFragment
+    private lateinit var creditCardButton: Button
+    private lateinit var giftCardButton: Button
+    private lateinit var cashButton: Button
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityMyCreditCartsBinding.inflate(layoutInflater)
+        binding = ActivitySelectPaymentMethodBinding.inflate(layoutInflater)
         val view = binding.root
         setContentView(view)
 
@@ -27,14 +31,23 @@ class MyCreditCards : AppCompatActivity(), NavigationView.OnNavigationItemSelect
         this.setSupportActionBar(appBar)
         setupDrawer(appBar)
 
-        creditCardFragment = myCreditCardsFragment()
-        binding.cardItem.selectCard.setOnClickListener {
-            val intent = Intent(this, SplashScreenProcessingPayment::class.java)
+        creditCardButton = findViewById(R.id.creditCardPaymentButton)
+        giftCardButton = findViewById(R.id.giftCardPaymentButton)
+        cashButton = findViewById(R.id.cashPaymentButton)
+ //       acceptPaymentButton = findViewById(R.id.acceptPaymentButton)
+
+        creditCardButton.setOnClickListener() {
+            val intent = Intent(this, MyCreditCards::class.java)
             this.startActivity(intent)
         }
 
-        binding.btnAgregar.setOnClickListener {
-            val intent = Intent(this, CreditCardActivity::class.java)
+        giftCardButton.setOnClickListener() {
+            val intent = Intent(this, MyGiftCards::class.java)
+            this.startActivity(intent)
+        }
+
+        cashButton.setOnClickListener() {
+            val intent = Intent(this, CashPaymentActivity::class.java)
             this.startActivity(intent)
         }
 
@@ -45,11 +58,14 @@ class MyCreditCards : AppCompatActivity(), NavigationView.OnNavigationItemSelect
 
         binding.navView.setNavigationItemSelectedListener(this)
     }
-
     private fun setupDrawer(toolbar: Toolbar){
         val drawerLayout = findViewById<DrawerLayout>(R.id.drawer_layout)
-        val drawerToggle = ActionBarDrawerToggle(this,drawerLayout,toolbar,R.string.open_drawer,R.string.close_drawer)
+        val drawerToggle = ActionBarDrawerToggle(this,drawerLayout,toolbar,
+            R.string.open_drawer,
+            R.string.close_drawer
+        )
     }
+
     override fun onNavigationItemSelected(item: MenuItem) : Boolean {
         // Handle navigation view item clicks here.
