+++ app/src/main/java/com/example/motomoapp/view/myGiftCardsFragment.kt
@@ -1,0 +1,4 @@
+package com.example.motomoapp.view
+
+class myGiftCardsFragment {
+}
