+++ app/src/main/res/layout/activity_cart_summary.xml
@@ -7,7 +7,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CartSummaryActivity"
+    tools:context=".view.CartSummaryActivity"
     >
 
     <include layout="@layout/app_bar"
@@ -18,7 +18,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".ItemDetalleActivity">
+        tools:context=".view.ItemDetalleActivity">
 
         <com.google.android.material.button.MaterialButton
             android:id="@+id/btnBack"
