+++ app/src/main/res/layout/activity_cash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/white"
     android:layout_height="match_parent"
-    tools:context=".CashActivity">
+    tools:context=".view.CashActivity">
 
     <include
         android:id="@+id/include"
