+++ app/src/main/res/layout/activity_cash_payment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".CashPaymentActivity">
+    tools:context=".view.CashPaymentActivity">
 
     <ImageView
         android:id="@+id/done"
