+++ app/src/main/res/layout/activity_credit_card.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context="CreditCardActivity">
+    tools:context=".view.CreditCardActivity">
 
     <!-- Campos del formulario para tarjeta de crédito -->
     <include
