+++ app/src/main/res/layout/activity_gift_card.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".GiftCardActivity">
+    tools:context=".view.GiftCardActivity">
 
 
     <include
