+++ app/src/main/res/layout/activity_gps.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".GpsActivity">
+        tools:context=".view.GpsActivity">
 
     <include layout="@layout/app_bar_tabs"
         android:layout_width="match_parent"
