+++ app/src/main/res/layout/activity_menu_inicio.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".MenuInicioActivity"
+    tools:context=".view.inicio.MenuInicioActivity"
     >
 
     <include layout="@layout/app_bar"
