+++ app/src/main/res/layout/activity_my_credit_carts.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".MyCreditCards"
+    tools:context=".view.MyCreditCards"
     >
 
     <include layout="@layout/app_bar"
@@ -20,7 +20,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MyCreditCards">
+        tools:context=".view.MyCreditCards">
 
         <TextView
             android:id="@+id/myCreditCards"
