+++ app/src/main/res/layout/activity_order.xml
@@ -9,7 +9,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
 
-    tools:context=".OrderActivity"
+    tools:context=".view.OrderActivity"
     >
 
     <include layout="@layout/app_bar_tabs"
