+++ app/src/main/res/layout/activity_payment_done.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".PaymentDone">
+    tools:context=".view.PaymentDone">
 
     <!--
     <include layout="@layout/app_bar"
