+++ app/src/main/res/layout/activity_payment_method.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".PaymentMethodActivity">
+    tools:context=".view.PaymentMethodActivity">
 
     <include
         android:id="@+id/include"
