+++ app/src/main/res/layout/activity_select_payment_method.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:background="@color/white"
     android:layout_height="match_parent"
-    tools:context=".SelectPaymentMethodActivity"
+    tools:context=".view.SelectPaymentMethodActivity"
     >
 
     <include layout="@layout/app_bar"
