+++ app/src/main/res/layout/activity_splash_screen_processing_payment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SplashScreenProcessingPayment">
+    tools:context=".view.SplashScreenProcessingPayment">
 
     <ImageView
         android:id="@+id/processingPaymentImg"
