+++ app/src/main/res/layout/fragment_cards_item_list.xml
@@ -10,6 +10,6 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".myCreditCardsFragment"
+    tools:context=".view.myCreditCardsFragment"
     tools:listitem="@layout/fragment_card_item"
 />
