+++ app/src/main/res/layout/fragment_giftcards_item_list.xml
@@ -10,6 +10,6 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".myGiftCardsFragment"
+    tools:context=".view.myGiftCardsFragment"
     tools:listitem="@layout/fragment_giftcard_item"
     />
