+++ app/src/main/res/layout/items_grid.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".GridFragment">
+    tools:context=".view.GridFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerProducts"
